requirements.txt
requests
vcrpy #records http responses during tests
pytest